EditAndContinueDocumentAnalysesCache
